analysis/pyproject.toml
[project]
name = "analysis"
version = "0.1.0"
description = ""
authors = [
    {name = "graxel"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "numpy (>=2.3.1,<3.0.0)",
    "plotly (>=6.2.0,<7.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "notebook (>=7.4.4,<8.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "pywavelets (>=1.8.0,<2.0.0)",
    "scipy (>=1.16.0,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
